+++ src/domain/Entities/ServiceEntity.cs
@@ -3,13 +3,13 @@ namespace Domain.Entities
     public class ServiceEntity : AbstractEntity
     {
         public string Name { get; set; }
-        public uint Cost { get; set; }
+        public int Cost { get; set; }
 
         public ServiceEntity(string name, uint cost)
         {
             Identify();
             Name = name;
-            Cost = cost;
+            Cost = (int) cost;
         }
     }
 }
